added microbit .hex file
for testing the setup.

connect pin0 and Ground pin of your micro:bit to a buzzer.
load the .hex file on to your micro:bit.
on the website connect to your micro:bit, type "M1:", "M2:" or "M3:" in the input field and click send.
You should hear the robot play a melody.